compilation fix again 3

+++ backend/scalestore/rdma/MessageHandler.cpp
@@ -523,7 +523,7 @@ try_shuffle:
             latencyMeasureResults[aggregatedTimeMeasureCounter] = double(std::chrono::duration_cast<std::chrono::microseconds>(afterMessage - beforeMessage).count());
             aggregatedTimeMeasureCounter++;
             if(aggregatedTimeMeasureCounter == aggregatedMsgAmount){
-                for(int i = 0; i < aggregatedTimeMeasureCounter; i++){
+                for(uint64_t i = 0; i < aggregatedTimeMeasureCounter; i++){
                     aggregatedTimeMeasure += latencyMeasureResults[i];
                     std::cout<< latencyMeasureResults[i] << " ";
                 }
